requirements.txt
Django==5.2
djangorestframework==3.16.0
celery==5.5
redis==5.1.1
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.1.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
gunicorn==23.0.0
twilio==9.7.0
requests==2.32.5
python-dateutil==2.9.0.post0
whitenoise==6.6.0
django-timezone-field==7.1
django-recurrence==1.11.1
dj-database-url==3.0.1